local storage fix

+++ js/signup.js
@@ -1,5 +1,4 @@
-let inputUsername = document.getElementById("username").value;
-let inputPassword = document.getElementById("password").value;
+
 
 class user {
     constructor (username, password, content) {
@@ -9,20 +8,33 @@ class user {
     }
 }
 
-const newUser = new user(inputUsername, inputPassword, []);
+
 
 function onRegister(event) {
     event.preventDefault();
+    let inputUsername = document.getElementById("username").value;
+    let inputPassword = document.getElementById("password").value;
+    let  usersArray= JSON.parse(localStorage.getItem("usersArray"));
+    const newUser = new user(inputUsername, inputPassword, []);
+    console.log('newUser: ', newUser);
+   
+
     if (usersArray === null) {
         localStorage.setItem("usersArray", JSON.stringify([newUser]));
     } else {
         let checkUsernameExists = usersArray.find(user => {
             if (user.username === inputUsername) {
                 alert("This username exists");
+               return true
             }
         })
+        if(checkUsernameExists){return;}
         localStorage.setItem("usersArray", JSON.stringify([...usersArray, newUser]));
         alert("You're signed up!")
     }
 }
+
+// function addNewUserObj(){
+    
+// }
